+++ src/Entities/Payments/PreCheckoutQuery.php
@@ -37,7 +37,7 @@ class PreCheckoutQuery extends Entity
     public function subEntities()
     {
         return [
-            'user'       => User::class,
+            'from'       => User::class,
             'order_info' => OrderInfo::class,
         ];
     }